// src/app/enums/months-of-the-year.enum.ts
export enum MonthsOfTheYear {
  JAN = 1,
  FEB = 2,
  MAR = 3,
  APR = 4,
  MAY = 5,
  JUN = 6,
  JUL = 7,
  AUG = 8,
  SEP = 9,
  OCT = 10,
  NOV = 11,
  DEC = 12
}
export const MONTHS_OPTIONS = [
  { label: 'Tháng 1', value: MonthsOfTheYear.JAN },
  { label: 'Tháng 2', value: MonthsOfTheYear.FEB },
  { label: 'Tháng 3', value: MonthsOfTheYear.MAR },
  { label: 'Tháng 4', value: MonthsOfTheYear.APR },
  { label: 'Tháng 5', value: MonthsOfTheYear.MAY },
  { label: 'Tháng 6', value: MonthsOfTheYear.JUN },
  { label: 'Tháng 7', value: MonthsOfTheYear.JUL },
  { label: 'Tháng 8', value: MonthsOfTheYear.AUG },
  { label: 'Tháng 9', value: MonthsOfTheYear.SEP },
  { label: 'Tháng 10', value: MonthsOfTheYear.OCT },
  { label: 'Tháng 11', value: MonthsOfTheYear.NOV },
  { label: 'Tháng 12', value: MonthsOfTheYear.DEC }
];